Lots of internal rearrangement, for very little apparent change: * Moved a lot of file opening logic out of Editor and Location, into File * Moved a lot of the document change tracking logic out of Editor, into File. * Renamed the File class to BaseFile; it seems File is a reserved name and was confusing C++ * Fixed a crash-bug when you used the file-open dialog a second time to open a second batch of files.
Why did I move more logic to BaseFile? I want it to become the central unit of fileness. I want to be able to create multiple Editors at the one File object, and have it Just Work. (Which it should in theory now)... That allows us to have split panes later easier.