Progress on the "Unsaved changes" dialog.
Also, some rejiggering of the whole OpenFileModel thing. Now there is an OpenFileManager, which just handles which files are open and their state without doing anything to look good in a tree view. 

OpenFileModel is going to get renamed and repurposed to being a treeview faceplate for OpenFileManager, to allow filtering. That'll let us reuse the components for the "Unsaved changes" dialog.