You can now put * in the path and it will attempt to open a sudo channel and crash in the process.
Why is this being committed if that crashes? Because if you put your password into sshconnection.cpp (line 224) it will actually work. Gotta make it intelligent enough to request the password as required. 

Also, it's unlikely anyone will stumble upon it since it all runs as it always did until you put * in the path.