Regular Expression Tester.
See, I was innocently walking past my computer when I tripped up and landed on the keyboard. When I looked up, the RegExp tester was done. It happens to me sometimes. 

In particular, try selecting some text before opening the RegExp tester. It auto fills the regexp field for you.