Start of code to detect changes to remote files.
Current status: 
Server polls files every few seconds for changes, sends a notification down the line to the client about it. Client has the smarts to detect the differences between a server notification and a request-response, logs notification messages to the qDebug() console. Doesn't act on them yet. 

However, there is a big limitation. Currently, all comms between client and server are initiated by the client side. Notifications from slave are not seen by the client until it sends a message to the server. 

Need to find a way to support server-side initiated comms before this feature can be finished. (Currently in a state where it's harmless, and can be committed to trunk though it's largely useless)